+++ test/assent/http_adapter/mint_test.exs
@@ -2,6 +2,7 @@ defmodule Assent.HTTPAdapter.MintTest do
   use ExUnit.Case
   doctest Assent.HTTPAdapter.Mint
 
+  alias ExUnit.CaptureIO
   alias Mint.TransportError
   alias Assent.HTTPAdapter.{HTTPResponse, Mint}
 
@@ -12,7 +13,9 @@ defmodule Assent.HTTPAdapter.MintTest do
 
       mint_opts = [transport_opts: [cacerts: TestServer.x509_suite().cacerts], protocols: [:http1]]
 
-      assert {:ok, %HTTPResponse{status: 200, body: "HTTP/1.1"}} = Mint.request(:get, TestServer.url(), nil, [], mint_opts)
+      assert CaptureIO.capture_io(:stderr, fn ->
+        assert {:ok, %HTTPResponse{status: 200, body: "HTTP/1.1"}} = Mint.request(:get, TestServer.url(), nil, [], mint_opts)
+      end) =~ "Assent.HTTPAdapter.Mint is deprecated, consider use Assent.HTTPAdapter.Finch instead"
     end
 
     test "handles SSL with bad certificate" do
@@ -21,7 +24,9 @@ defmodule Assent.HTTPAdapter.MintTest do
       bad_host_url = TestServer.url(host: "bad-host.localhost")
       mint_opts = [transport_opts: [cacerts: TestServer.x509_suite().cacerts]]
 
-      assert {:error, %TransportError{reason: {:tls_alert, {:handshake_failure, _error}}}} = Mint.request(:get, bad_host_url, nil, [], mint_opts)
+      assert CaptureIO.capture_io(:stderr, fn ->
+        assert {:error, %TransportError{reason: {:tls_alert, {:handshake_failure, _error}}}} = Mint.request(:get, bad_host_url, nil, [], mint_opts)
+      end) =~ "Assent.HTTPAdapter.Mint is deprecated, consider use Assent.HTTPAdapter.Finch instead"
     end
 
     test "handles SSL with bad certificate and no verification" do
@@ -31,7 +36,9 @@ defmodule Assent.HTTPAdapter.MintTest do
       bad_host_url = TestServer.url(host: "bad-host.localhost")
       mint_opts = [transport_opts: [cacerts: TestServer.x509_suite().cacerts, verify: :verify_none]]
 
-      assert {:ok, %HTTPResponse{status: 200}} = Mint.request(:get, bad_host_url, nil, [], mint_opts)
+      assert CaptureIO.capture_io(:stderr, fn ->
+        assert {:ok, %HTTPResponse{status: 200}} = Mint.request(:get, bad_host_url, nil, [], mint_opts)
+      end) =~ "Assent.HTTPAdapter.Mint is deprecated, consider use Assent.HTTPAdapter.Finch instead"
     end
 
     if :crypto.supports()[:curves] do
@@ -41,7 +48,9 @@ defmodule Assent.HTTPAdapter.MintTest do
 
         mint_opts = [transport_opts: [cacerts: TestServer.x509_suite().cacerts]]
 
-        assert {:ok, %HTTPResponse{status: 200, body: "HTTP/2"}} = Mint.request(:get, TestServer.url(), nil, [], mint_opts)
+        assert CaptureIO.capture_io(:stderr, fn ->
+          assert {:ok, %HTTPResponse{status: 200, body: "HTTP/2"}} = Mint.request(:get, TestServer.url(), nil, [], mint_opts)
+        end) =~ "Assent.HTTPAdapter.Mint is deprecated, consider use Assent.HTTPAdapter.Finch instead"
       end
     else
       IO.warn("No support curve algorithms, can't test in #{__MODULE__}")
@@ -52,7 +61,9 @@ defmodule Assent.HTTPAdapter.MintTest do
       url = TestServer.url()
       TestServer.stop()
 
-      assert {:error, %TransportError{reason: :econnrefused}} = Mint.request(:get, url, nil, [])
+      assert CaptureIO.capture_io(:stderr, fn ->
+        assert {:error, %TransportError{reason: :econnrefused}} = Mint.request(:get, url, nil, [])
+      end) =~ "Assent.HTTPAdapter.Mint is deprecated, consider use Assent.HTTPAdapter.Finch instead"
     end
 
     test "handles query in URL" do
@@ -62,7 +73,9 @@ defmodule Assent.HTTPAdapter.MintTest do
         Plug.Conn.send_resp(conn, 200, "")
       end)
 
-      assert {:ok, %HTTPResponse{status: 200}} = Mint.request(:get, TestServer.url("/get?a=1"), nil, [])
+      assert CaptureIO.capture_io(:stderr, fn ->
+        assert {:ok, %HTTPResponse{status: 200}} = Mint.request(:get, TestServer.url("/get?a=1"), nil, [])
+      end) =~ "Assent.HTTPAdapter.Mint is deprecated, consider use Assent.HTTPAdapter.Finch instead"
     end
 
     test "handles POST" do
@@ -77,7 +90,9 @@ defmodule Assent.HTTPAdapter.MintTest do
         Plug.Conn.send_resp(conn, 200, "")
       end)
 
-      assert {:ok, %HTTPResponse{status: 200}} = Mint.request(:post, TestServer.url("/post"), "a=1&b=2", [{"content-type", "application/x-www-form-urlencoded"}])
+      assert CaptureIO.capture_io(:stderr, fn ->
+        assert {:ok, %HTTPResponse{status: 200}} = Mint.request(:post, TestServer.url("/post"), "a=1&b=2", [{"content-type", "application/x-www-form-urlencoded"}])
+      end) =~ "Assent.HTTPAdapter.Mint is deprecated, consider use Assent.HTTPAdapter.Finch instead"
     end
   end
 end