+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.2.6 (TBA)
 
 * Added `Assent.HTTPAdapter.Finch`
+* Deprecated `Assent.HTTPAdapter.Mint`
 
 ## v0.2.5 (2023-08-21)